# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Elektra Endpoints",
        "description": "Elektra Endpoints",
        "termsOfService": "http://programming4food.io/terms/",
        "contact": {
            "email": "developers@programming4food.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/pedidos": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Listado de Pedidos",
                "description": "Este endpoint retorna los pedidos",
                "operationId": "PedidosList",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pedido"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Couldn't Create Token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/pedidos/{pedidos_id}": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Detalle del pedido",
                "description": "Este endpoint retorna los pedidos",
                "operationId": "Pedido",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pedido"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Couldn't Create Token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/pedidos/{pedidos_id}/tracking": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Detalle de envio del pedido",
                "description": "Este endpoint retorna el estado de envio de un pedido",
                "operationId": "PedidoTracking",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pedido"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Couldn't Create Token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/pedidos/{pedidos_id}/items/{product_id}/review": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Detalle de envio del pedido",
                "description": "Este endpoint retorna el estado de envio de un pedido",
                "operationId": "PedidoTracking",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pedido"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Couldn't Create Token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Listado de Productos",
                "description": "Este endpoint retorna los productos del catalogo de Elektra que vienen del Dataset",
                "operationId": "ProductsList",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pedido"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Couldn't Create Token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Error": {
                "properties": {
                    "error": {
                        "description": "Mensaje de Error",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Pedido": {
                "properties": {
                    "id": {
                        "description": "Id de Pedido",
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        }
    }
}